+++ app/src/main/kotlin/org/strigate/ferrot/presentation/screen/UpdatesScreen.kt
@@ -28,12 +28,15 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import org.strigate.ferrot.R
+import org.strigate.ferrot.extensions.toast
 import org.strigate.ferrot.presentation.component.settings.StaticSettingsSection
 import org.strigate.ferrot.presentation.component.settings.SwitchSetting
 import org.strigate.ferrot.presentation.component.settings.TextSetting
 import org.strigate.ferrot.presentation.component.state.ErrorState
 import org.strigate.ferrot.presentation.component.state.LoadingState
+import org.strigate.ferrot.presentation.event.UpdatesEvent
 import org.strigate.ferrot.presentation.state.UpdatesUiState
+import org.strigate.ferrot.presentation.theme.FerrotTopAppBarDefaults
 import org.strigate.ferrot.presentation.theme.LocalDimens
 import org.strigate.ferrot.presentation.util.UiFormatter
 import org.strigate.ferrot.presentation.viewmodel.UpdatesViewModel
@@ -52,10 +55,20 @@ fun UpdatesScreen(
     LaunchedEffect(Unit) {
         viewModel.logShown()
     }
+    LaunchedEffect(Unit) {
+        viewModel.event.collect { event ->
+            when (event) {
+                is UpdatesEvent.ShowToast -> {
+                    context.toast(event.textRes)
+                }
+            }
+        }
+    }
 
     Scaffold(
         topBar = {
             TopAppBar(
+                colors = FerrotTopAppBarDefaults.colors(),
                 navigationIcon = {
                     IconButton(
                         onClick = { backDispatcher?.onBackPressed() },
@@ -76,7 +89,7 @@ fun UpdatesScreen(
                 modifier = modifier
                     .fillMaxSize()
                     .padding(contentPadding),
-                color = MaterialTheme.colorScheme.surface,
+                color = MaterialTheme.colorScheme.background,
             ) {
                 when (val state = uiState) {
                     is UpdatesUiState.Loading -> {