+++ app/src/main/kotlin/org/strigate/ferrot/work/UpdateDependenciesWorker.kt
@@ -48,11 +48,11 @@ class UpdateDependenciesWorker(
             if (isAppInForeground()) {
                 when (updateStatus) {
                     YoutubeDL.UpdateStatus.ALREADY_UP_TO_DATE -> {
-                        appContext.toast(appContext.getString(R.string.toast_dependencies_already_up_to_date))
+                        appContext.toast(appContext.getString(R.string.toast_no_dependency_updates_available))
                     }
 
                     YoutubeDL.UpdateStatus.DONE -> {
-                        appContext.toast(appContext.getString(R.string.toast_dependencies_update_complete))
+                        appContext.toast(appContext.getString(R.string.toast_dependencies_updated))
                     }
 
                     else -> {